pyproject.toml
[tool.poetry]
name = "vigilant-octo-bassoon"
version = "0.1.0"
description = ""
authors = ["Derli Machado <dam.machado1@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.6"
djangorestframework = "^3.12.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
